requirements.txt
asgiref==3.5.0
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-on-heroku==1.1.2
django-widget-tweaks==1.4.12
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
Werkzeug==2.0.3
whitenoise==6.0.0
wincertstore==0.2
